+++ src/modules/IntexDeviceCommand.ts
@@ -1,7 +1,7 @@
 /**
  * Type class representing an Intex command definition.
  */
-export class IntexDeviceCommandDefinition {
+export class IntexDeviceCommand {
   id: number | undefined;
   commandSetTypeId: number | undefined;
   currentVersion: string | undefined;