+++ src/modules/IntexDeviceCommandResult.ts
@@ -1,0 +1,7 @@
+/**
+ * Type class representing an Intex device command result.
+ */
+export class IntexDeviceCommandResult {
+  code: number | undefined;
+  msg: string | undefined;
+}