+++ src/modules/IntexDeviceCommandStatus.ts
@@ -1,0 +1,9 @@
+/**
+ * Type class representing an Intex device command status.
+ */
+export class IntexDeviceCommandStatus {
+  c2DCommandStatus: string | undefined;
+  deviceId: string | undefined;
+  iotDataType: string | undefined;
+  sessionId: string | undefined;
+}