+++ src/nodes/intexspa-action/intexspa-action.html/modules/types.ts
@@ -1,0 +1,4 @@
+import { EditorNodeProperties } from 'node-red';
+import { IntexSpaActionOptions } from '../../shared/types';
+
+export interface IntexSpaActionEditorNodeProperties extends EditorNodeProperties, IntexSpaActionOptions {}