+++ src/nodes/intexspa-action/shared/types.ts
@@ -1,0 +1,4 @@
+export interface IntexSpaActionOptions {
+  device: string;
+  component: string;
+}