+++ src/nodes/intexspa-config/modules/types.ts
@@ -6,4 +6,5 @@ export interface IntexSpaConfigNodeDef extends NodeDef, IntexSpaConfigOptions {}
 export interface IntexSpaConfigNode extends Node {
   username: string;
   password: string;
+  device: string;
 }