+++ src/nodes/intexspa-config/shared/types.ts
@@ -1,4 +1,5 @@
 export interface IntexSpaConfigOptions {
   username: string;
   password: string;
+  device: string;
 }